Template overload of <size_t> was incorrect
size_t can be 32-bit or 64-bit, was hardcoded to 32-bit

+++ llvm/_core.cpp
@@ -63,7 +63,7 @@ template <> PyObject* pycap_new<int>(int i)
     return PyLong_FromLong(static_cast<long>(i));
 }
 
-template <> PyObject* pycap_new<size_t>(size_t i)
+template <> PyObject* pycap_new<unsigned long>(unsigned long i)
 {
     return PyLong_FromLong(static_cast<unsigned long>(i));
 }
